+++ RefactorMCP.ConsoleApp/SyntaxWalkers/InstanceMemberNameWalker.cs
@@ -7,12 +7,6 @@ namespace RefactorMCP.ConsoleApp.SyntaxWalkers
     {
         public override void VisitFieldDeclaration(FieldDeclarationSyntax node)
         {
-            if (node.Modifiers.Any(SyntaxKind.StaticKeyword))
-            {
-                base.VisitFieldDeclaration(node);
-                return;
-            }
-
             foreach (var variable in node.Declaration.Variables)
                 Add(variable.Identifier.ValueText);
             base.VisitFieldDeclaration(node);